Interfaces implemented, Dust particles, pistol
3 main interfaces, iDamageable, iCollectible and iInteractible, mainly using the iDamageable for now, and the collectible for the ammo.
Dust particles are almost done, nelson sent me more stuff, i will try to end it tomorrow
Changed pistol mag mechanism because data structures professor said i was over engenering?, and he was right, now its simpler

+++ Assets/Scripts/Enemies/MutantTrees/Attack/MutantTreeAttack.cs
@@ -13,8 +13,7 @@ public class MutantTreeAttack : MonoBehaviour
     [SerializeField] private GameObject sap;
     [SerializeField] private GameObject sapHolder;
     [SerializeField] private GameObject sapOrigin;
-    [SerializeField] private int shootDelaySeconds;
-
+    [SerializeField] private float shootDelaySeconds;
 
     void Update()
     {
